+++ packages/4c-file-butler/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@4c/file-butler",
   "version": "1.1.1",
-  "main": "index.js",
+  "main": "lib.js",
   "repository": {
     "type": "git",
     "url": "https://github.com/4Catalyzer/file-butler.git"
@@ -12,13 +12,16 @@
     "tdd": "jest --watch",
     "test": "npm run lint",
     "testonly": "jest",
-    "lint": "eslint . && prettier --list-different --ignore-path .eslintignore '**/*.{json,css,md}'",
-    "format": "eslint . --fix && prettier --write --ignore-path .eslintignore '**/*.{json,css,md}'",
+    "lint":
+      "eslint . && prettier --list-different --ignore-path .eslintignore '**/*.{json,css,md}'",
+    "format":
+      "eslint . --fix && prettier --write --ignore-path .eslintignore '**/*.{json,css,md}'",
     "precommit": "lint-staged"
   },
   "bin": {
     "file-butler": "index.js"
   },
+  "files": ["lib.js", "index.js"],
   "engines": {
     "node": ">= 8.3"
   },
@@ -31,25 +34,18 @@
     "trailingComma": "all"
   },
   "lint-staged": {
-    "*.js": [
-      "eslint --fix",
-      "git add"
-    ],
+    "*.js": ["eslint --fix", "git add"],
     "*.{json,css,md}": [
       "prettier --write --ignore-path .eslintignore",
       "git add"
     ]
   },
   "jest": {
-    "roots": [
-      "<rootDir>/test"
-    ],
+    "roots": ["<rootDir>/test"],
     "testEnvironment": "node"
   },
   "release": {
-    "extends": [
-      "@4c/semantic-release-config"
-    ]
+    "extends": ["@4c/semantic-release-config"]
   },
   "devDependencies": {
     "@4c/semantic-release-config": "^1.0.2",
